requirements/base.txt
celery==4.0.0
dj-database-url==0.4.1
Django==1.11.29
django-bootstrap-pagination==1.6.2
django-bootstrap3==11.0.0
django-braces==1.11.0
django-celery-beat==1.0.0
django-filter==1.1.0
django-mptt==0.8.6
django-mysql==1.1.0
django-pandas==0.4.1
django-pipeline==1.6.9
django-redis==4.7.0
django-selectable==1.2.1
django-sendfile==0.3.11
django-subdomains==2.1.0
djangorestframework==3.9.4
drf-yasg==1.17.1
fuzzywuzzy==0.12.0
mysqlclient==1.3.7
networkx==1.11
numexpr==2.6.1
pandas==0.18.1
Parsley==1.3
prettyconf==1.2.3
pyparsing==2.1.9
python-dateutil==2.5.3
python-Levenshtein==0.12.0
RapidSMS==1.0.0
raven==6.10.0
redis==2.10.5
sqlparse==0.3.0
tablib==0.13.0
uWSGI==2.0.18
whitenoise==4.1.3
Bottleneck==1.2.0
